move test to the new package
