+++ components/movementsensor/rtkutils/cachedData_test.go
@@ -1,0 +1,49 @@
+package rtkutils
+
+import (
+	"context"
+	"testing"
+
+	geo "github.com/kellydunn/golang-geo"
+	"go.viam.com/test"
+)
+
+var loc = geo.NewPoint(90, 1)
+
+const (
+	alt        = 50.5
+	speed      = 5.4
+	activeSats = 1
+	totalSats  = 2
+	hAcc       = 0.7
+	vAcc       = 0.8
+	fix        = 1
+)
+
+func TestReadingsSerial(t *testing.T) {
+	ctx := context.Background()
+	g := NewCachedGpsData()
+	g.uncachedData = GPSData{
+		Location:   loc,
+		Alt:        alt,
+		Speed:      speed,
+		VDOP:       vAcc,
+		HDOP:       hAcc,
+		SatsInView: totalSats,
+		SatsInUse:  activeSats,
+		FixQuality: fix,
+	}
+
+	loc1, alt1, err := g.Position(ctx, make(map[string]interface{}))
+	test.That(t, err, test.ShouldBeNil)
+	test.That(t, loc1, test.ShouldEqual, loc)
+	test.That(t, alt1, test.ShouldEqual, alt)
+
+	speed1, err := g.LinearVelocity(ctx, make(map[string]interface{}))
+	test.That(t, err, test.ShouldBeNil)
+	test.That(t, speed1.Y, test.ShouldEqual, speed)
+
+	fix1, err := g.ReadFix(ctx)
+	test.That(t, err, test.ShouldBeNil)
+	test.That(t, fix1, test.ShouldEqual, fix)
+}
